refactor(structure): consolidate workers, i18n barrel, and Rust test organization
Frontend:
- Move ThumbnailWorkerPool.ts from src/lib/workers/ → src/workers/
- Add barrel export at src/workers/index.ts
- Remove src/lib/workers/
- Update import sites: performanceAdapter.ts, transport.ts
- Add src/i18n/index.ts barrel export (I18nProvider, useI18n)

Rust backend:
- Move root-level test files out of src-tauri/src/:
    decoder_pool_stress_test.rs  → thumbnail_engine/stress_test.rs
    thumbnail_engine_tests.rs    → thumbnail_engine/tests.rs
    thumbnail_engine_proptest.rs → thumbnail_engine/proptest.rs
- Declare submodules in thumbnail_engine.rs
- Clean up lib.rs module declarations

All verification passing:
  TypeScript: 0 errors
  Vitest: 141 files / 1452 tests passed
  Rust cargo test: 81/81 passed
