+++ src-tauri/src/thumbnail_engine/stress_test.rs
