+++ src/i18n/index.ts
@@ -1,0 +1,1 @@
+export { I18nProvider, useI18n } from "./I18nProvider";
