+++ src/workers/ThumbnailWorkerPool.ts
