+++ src/workers/index.ts
@@ -1,0 +1,1 @@
+export { ThumbnailWorkerPool } from "./ThumbnailWorkerPool";
